+++ R/raw.getSampleFiles.R
@@ -1,6 +1,5 @@
 #' provide sample RAW data files
 #' @returns path and filename with sample data
-#' @export
 raw.getSampleFiles <- function() {
   tmpDir = tempdir()
   filename = file.path(tmpDir,"20241001_spinPc_TG_text_Sample.txt")